backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# 任务队列
celery==5.3.4
redis==5.0.1

# 向量数据库
chromadb==1.0.16

# AI处理
openai==1.3.8

# 工具库
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1